Moved distanceBetween to newly created gameFunctions.lua
