+++ gameFunctions.lua
@@ -1,0 +1,3 @@
+function distanceBetween(x1, y1, x2, y2)
+    return math.sqrt((x2 - x1)^2 + (y2 - y1)^2)
+end
