+++ gameVariables.lua
